remove monitoring dashboard legacy code and corresponding code, minor fixes
