+++ pgmanage/app/static/assets/js/pgmanage_frontend/src/tree_context_functions/tree_mariadb.js
@@ -483,8 +483,7 @@ function getTreeDetailsMariadb(node) {
                         text: 'Dashboard',
                         icon: 'fas cm-all fa-chart-line',
                         action: function(node) {
-                            v_connTabControl.tag.createMonitorDashboardTab();
-                            startMonitorDashboard();
+                            v_connTabControl.tag.createMonitoringDashboardTab();
                         }
                     }, */{
                         text: 'Process List',
@@ -524,8 +523,8 @@ function getTreeDetailsMariadb(node) {
 
             if (v_connTabControl.selectedTab.tag.firstTimeOpen) {
               v_connTabControl.selectedTab.tag.firstTimeOpen = false;
-              //v_connTabControl.tag.createMonitorDashboardTab();
-              //startMonitorDashboard();
+              //v_connTabControl.tag.createMonitoringDashboardTab();
+
             }
