+++ pgmanage/app/static/assets/js/pgmanage_frontend/src/tree_context_functions/tree_mysql.js
@@ -477,8 +477,7 @@ function getTreeDetailsMysql(node) {
                         text: 'Dashboard',
                         icon: 'fas cm-all fa-chart-line',
                         action: function(node) {
-                            v_connTabControl.tag.createMonitorDashboardTab();
-                            startMonitorDashboard();
+                            v_connTabControl.tag.createMonitoringDashboardTab();
                         }
                     }, */ {
               text: "Process List",
@@ -531,8 +530,7 @@ function getTreeDetailsMysql(node) {
 
       if (v_connTabControl.selectedTab.tag.firstTimeOpen) {
         v_connTabControl.selectedTab.tag.firstTimeOpen = false;
-        //v_connTabControl.tag.createMonitorDashboardTab();
-        //startMonitorDashboard();
+        //v_connTabControl.tag.createMonitoringDashboardTab();
       }
     },
     function (p_return) {
