+++ pgmanage/app/templates/app/workspace.html
@@ -216,26 +216,6 @@
     </div>
 
 
-    <!-- Modal Monitoring Units -->
-    <div class="modal fade" id="modal_monitoring_units" tabindex="-1" role="dialog" aria-hidden="true">
-      <div class="modal-dialog modal-lg" role="document">
-        <div class="modal-content">
-          <div class="modal-header align-items-center">
-            <h2 class="modal-title font-weight-bold">Monitoring Units</h2>
-            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-              <span aria-hidden="true"><i class="fa-solid fa-xmark"></i></span>
-            </button>
-          </div>
-          <div class="modal-body">
-            <div id="monitoring_units_grid" class="tabulator-custom" style="width: 100%; height: 300px; overflow: hidden;"></div>
-          </div>
-          <div class="modal-footer">
-            <button id="modal_monitoring_units_new_unit_btn" class='btn btn-primary btn-sm mr-3'>New Unit</button>
-          </div>
-        </div>
-      </div>
-    </div>
-
     <!-- Modal Users -->
     <div class="modal fade" id="modal_users" tabindex="-1" role="dialog" aria-hidden="true">
       <div class="modal-dialog" role="document">
@@ -270,27 +250,6 @@
             </div>
 
             <div id="plugin_grid" style="width: 100%; height: 300px; overflow: hidden;"></div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <!-- Modal Monitoring Unit Test -->
-    <div class="modal fade" id="modal_monitoring_unit_test" tabindex="-1" role="dialog" aria-hidden="true">
-      <div class="modal-dialog" role="document">
-        <div class="modal-content">
-          <div class="modal-header align-items-center">
-            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-              <span aria-hidden="true"><i class="fa-solid fa-xmark"></i></span>
-            </button>
-          </div>
-          <div class="modal-body">
-            <div class="dashboard_unit_content_group">
-              <div id="monitoring_unit_test_result">
-              </div>
-              <div id="monitoring_unit_test_legend" class="dashboard_unit_legend_box">
-              </div>
-            </div>
           </div>
         </div>
       </div>
