+++ pgmanage/app/urls.py
@@ -294,26 +294,16 @@ base_urlpatterns = [
     re_path(r'^get_table_definition_sqlite/', views.tree_sqlite.get_table_definition, name="get_table_definition_sqlite"),
 
     #MONITORING SYSTEM
-    path("monitoring-widgets", views.monitor_dashboard.monitoring_widgets, name="monitoring_widgets"),
-    path("monitoring-widgets/list", views.monitor_dashboard.monitoring_widgets_list,),
-    path("monitoring-widgets/test", views.monitor_dashboard.test_monitoring_widget),
-    path("monitoring-widgets/create", views.monitor_dashboard.create_dashboard_monitoring_widget),
-    path("monitoring-widgets/<int:widget_id>", views.monitor_dashboard.widget_detail),
-    path("monitoring-widgets/<int:widget_id>/template", views.monitor_dashboard.widget_template,),
-    path("monitoring-widgets/<int:widget_saved_id>/refresh", views.monitor_dashboard.refresh_monitoring_widget),
-    path("monitoring-widgets/user-created", views.monitor_dashboard.create_widget),
-    path("monitoring-widgets/user-created/<int:widget_id>", views.monitor_dashboard.user_created_widget_detail,),
-    re_path(r'^test_monitor_script/', views.monitor_dashboard.test_monitor_script, name='test_monitor_script'),
-    re_path(r'^get_monitor_unit_list/', views.monitor_dashboard.get_monitor_unit_list, name='get_monitor_unit_list'),
-    re_path(r'^get_monitor_unit_details/', views.monitor_dashboard.get_monitor_unit_details, name='get_monitor_unit_details'),
-    re_path(r'^get_monitor_units/', views.monitor_dashboard.get_monitor_units, name='get_monitor_units'),
-    re_path(r'^refresh_monitor_units/', views.monitor_dashboard.refresh_monitor_units, name='refresh_monitor_units'),
-    re_path(r'^get_monitor_unit_template/', views.monitor_dashboard.get_monitor_unit_template, name='get_monitor_unit_template'),
-    re_path(r'^save_monitor_unit/', views.monitor_dashboard.save_monitor_unit, name='save_monitor_unit'),
-    re_path(r'^delete_monitor_unit/', views.monitor_dashboard.delete_monitor_unit, name='delete_monitor_unit'),
-    re_path(r'^remove_saved_monitor_unit/', views.monitor_dashboard.remove_saved_monitor_unit, name='remove_saved_monitor_unit'),
-    re_path(r'^update_saved_monitor_unit_interval/', views.monitor_dashboard.update_saved_monitor_unit_interval, name='update_saved_monitor_unit_interval'),
-
+    path("monitoring-widgets", views.monitoring_dashboard.monitoring_widgets,),
+    path("monitoring-widgets/list", views.monitoring_dashboard.monitoring_widgets_list,),
+    path("monitoring-widgets/test", views.monitoring_dashboard.test_monitoring_widget),
+    path("monitoring-widgets/create", views.monitoring_dashboard.create_dashboard_monitoring_widget),
+    path("monitoring-widgets/<int:widget_id>", views.monitoring_dashboard.widget_detail),
+    path("monitoring-widgets/<int:widget_id>/template", views.monitoring_dashboard.widget_template,),
+    path("monitoring-widgets/<int:widget_saved_id>/refresh", views.monitoring_dashboard.refresh_monitoring_widget),
+    path("monitoring-widgets/user-created", views.monitoring_dashboard.create_widget),
+    path("monitoring-widgets/user-created/<int:widget_id>", views.monitoring_dashboard.user_created_widget_detail,),
+    
     # Configuration
     path('configuration/<int:config_id>/', views.configuration.delete_config, name="delete_configuration"),
     re_path(r'^configuration/$', views.configuration.get_configuration, name='get_configuration'),
