+++ pgmanage/app/views/__init__.py
@@ -1,4 +1,4 @@
 from . import (backup, bgjob, commands_history, configuration, connections,
-               file_manager, login, monitor_dashboard, plugins, polling,
+               file_manager, login, monitoring_dashboard, plugins, polling,
                restore, tree_mariadb, tree_mysql, tree_oracle, pgextras,
                tree_postgresql, tree_snippets, tree_sqlite, users, workspace)
